A huge list of changes

+++ src/main/java/com/continuum/FastRandom.java
@@ -51,7 +51,7 @@ public class FastRandom {
 	 * @return Random value
 	 */
 	public double randomDouble() {
-		return randomInt() / ((double) Integer.MAX_VALUE - 1d);
+		return randomLong() / ((double) Long.MAX_VALUE - 1d);
 	}
 
 	/**
